fixes #20
For some reason, we cannot play music that is inside a directory in the latest version of the playsound module. So, I made it install the v1.2.2 and updated requirements and the issue got fixed.

The main purpose I want the music to be inside a directory is because, we can then let the user decide what music they want to play, if we add more than one song we shouldn't clutter up the main directory.

So, i'm really glad that this is fixed :)!